utils: Move Ci from core to test
The class is solely used by tests.
